fix: tests coverage
