+++ src/InternalRouter.test.ts
@@ -3,24 +3,6 @@ import InternalRouter from './InternalRouter'
 
 describe('InternalRouter', () => {
   describe('route', () => {
-    it('calls registerStorageProvider', async () => {
-      const registerStorageProvider = jest.fn().mockResolvedValue(undefined)
-      const router = new InternalRouter({
-        registerStorageProvider,
-      } as any as Ketch)
-      await router.route('registerStorageProvider')
-      return expect(registerStorageProvider).toHaveBeenCalled()
-    })
-
-    it('calls setConsent', async () => {
-      const setConsent = jest.fn().mockResolvedValue(undefined)
-      const router = new InternalRouter({
-        setConsent,
-      } as any as Ketch)
-      await router.route('setConsent')
-      return expect(setConsent).toHaveBeenCalled()
-    })
-
     it('calls experienceClosed', async () => {
       const experienceClosed = jest.fn().mockResolvedValue(undefined)
       const router = new InternalRouter({
@@ -37,6 +19,15 @@ describe('InternalRouter', () => {
       } as any as Ketch)
       await router.route('invokeRight')
       return expect(invokeRight).toHaveBeenCalled()
+    })
+
+    it('calls setProvisionalConsent', async () => {
+      const setProvisionalConsent = jest.fn().mockResolvedValue(undefined)
+      const router = new InternalRouter({
+        setProvisionalConsent,
+      } as any as Ketch)
+      await router.route('setProvisionalConsent')
+      return expect(setProvisionalConsent).toHaveBeenCalled()
     })
   })
 })
